requirement.txt
Djangae==0.9.8
Django==1.9.7
djangorestframework==3.6.4
PyYaml==3.11
google-api-python-client
oauth2client==1.5.0
uritemplate==3.0.0
xhtml2pdf
httplib2==0.8
requests==2.2.1
six==1.10.0
rsa==3.4.2
html5lib==1.0b8
reportlab==3.3.0
urllib3==1.7.1
webob
sendmail==0.9
pycrypto==2.6.1
openpyxl